Code efficiency + compatibility improvement
Instead of constantly having to re-grab one variable in the page (very stupid), it now does this once and saves it to a global variable that is constantly accessed (improving speed by about 2 seconds when predicting). However, this means that now if your page changes too drastically (any special extensions such as browser reader), it may attempt to access an xpath to a variable that does not exist anymore because it was only grabbed on code startup and not dynamically inside the predict function. I may revert this change later as it makes it way easier to patch by bloxflip devs.
Made the code more compatible with earlier python versions, should now be able to check what version you have and adjust code accordingly.
Some minor UI changes